requirements.txt
altgraph==0.17.4
CTkMessagebox==2.7
customtkinter==5.2.2
darkdetect==0.8.0
defusedxml==0.7.1
Deprecated==1.2.18
fonttools==4.58.5
fpdf2==2.8.3
lxml==6.0.0
numpy==2.3.1
packaging==25.0
pdf2image==1.17.0
pikepdf==9.10.0
pillow==11.2.1
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyMuPDF==1.26.4
pypdf==5.6.1
setuptools==80.9.0
wrapt==1.17.2
